feat(releases): adds headsOnly flag to release to toggle full channel… (#340)
* feat(releases): adds headsOnly flag to release to toggle full channel pulls

The headsOnly configuration, in terms of releases, is meant to enable only the latest
OCP versions from channels to be mirrored. It defaults to true. When set to false, the min
and the max OCP versions in the channel will be chosen and the shortest upgrade path between the two
will be calculated and mirrored.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* docs: updates examples to include new release workflows

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* docs: fixes comment in release example docs to reflect the examples properly

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat(releases): adds ability to specify min or max version for OCP release

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: removes validate unit tests for removed release verifications

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>