feat: changes catalog and OCP release heads only workflows to use implicit ranges
When merging the catalog, unintentional changes to the upgrade and pruned bundles
can be created in the target FBC generated by oc-mirror.

To remove the requirement to merge FBCs, package ranges are being used in metadata for
heads-only workflows to generate FBCs that are scoped per the configuration instead of
generating differential FBCs and merging with previous FBCs. The OCP releases heads only
workflow has been updated work in the same way. To achieve the same optimization in regards
to download images, all images in the image mapping during each run are validated
against the metadata to make sure they do not exist in PastAssociations.

Closes #377

BREAKING CHANGE: This change adds PlatformMetadata and modifies OperatorMetadata on the Metadata Spec.
It also removes the catalog merging workflow.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>