chore(api): modifies ImagesetConfigurationSpec to conform to Kubernetes API Guidelines
Introduce Platform and PlatformType to replace OCP with
a more generic option. PlatformType is used to distinguish
between OKD and OCP. The original solution was to use the channel
name, but a specified type is most likely more survivable and allows
for later additions.

Replaces Repo type with Repositories in Helm Spec

Adds MarshalJSON and UnmarshalJSON methods to ImageType
to write the string representation to the metadata.json instead
of the int.

Removes redundant AdditionalImage and BlockedImages types

BREAKING CHANGE: Removes OCP and Repo types and invalidates past metadata

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>