chore(api): modifies ImagesetConfigurationSpec to conform to Kubernet… (#397)
* chore(api): modifies ImagesetConfigurationSpec to conform to Kubernetes API Guidelines

Introduce Platform and PlatformType to replace OCP with
a more generic option. PlatformType is used to distinguish
between OKD and OCP. The original solution was to use the channel
name, but a specified type is most likely more survivable and allows
for later additions.

Replaces Repo type with Repositories in Helm Spec

Adds MarshalJSON and UnmarshalJSON methods to ImageType
to write the string representation to the metadata.json instead
of the int.

Removes redundant AdditionalImage and BlockedImages types

BREAKING CHANGE: Removes OCP and Repo types and invalidates past metadata

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* docs: updates imageset-config-ref.yaml with new keys

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: modifies comments in UnmarshalJSON and MarshalJSON methods in API for type clarification

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: defines Helm and Platform fields to be consistent with fields in the spec

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>