feat(mirror): adds methods to prune images during mirroring process
The current images are evaluated and compared against the previous downloads.
If an image that has be previously download is not in the current image mapping,
the image is schedule to be pruned from the registry (manifest only). Currently,
only image specified by digest are being deleted. This is done on a best-effort bases as
some registry do not allow deletions.

Closes #302

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>