fix: changes ConvertToAssociationSet methods to allow for shared child manifest (#401)
Currently, if two images with an index manifest contain the same child manifest,
the conversion to an AssociationSet set fails because child manifests records are being
deleted after processing to identify parent image manifests. This change adds the child manifest
to a visited map and checks if it exists instead to allow for shared child manifests.

Fixes #402

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>