fix: incorporates feedback from PR code review
Reduces code duplication in mirror.go

Adds comment to clarify nullLogger for containerd Registry

Adds a check when proccessing next bundles in case there
is not next bundle and the channel head does not exist. Unlikely,
but if truem results in a panic.

Changes catalog image build to remove the entire configs
directory from the base image before adding the FBC to avoid
possible FBC aggregation with previous yaml or json file.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>