refactor: heads only workflow to remove the requirement to merge FBCs (#387)
* feat: changes catalog and OCP release heads only workflows to use implicit ranges

When merging the catalog, unintentional changes to the upgrade and pruned bundles
can be created in the target FBC generated by oc-mirror.

To remove the requirement to merge FBCs, package ranges are being used in metadata for
heads-only workflows to generate FBCs that are scoped per the configuration instead of
generating differential FBCs and merging with previous FBCs. The OCP releases heads only
workflow has been updated work in the same way. To achieve the same optimization in regards
to download images, all images in the image mapping during each run are validated
against the metadata to make sure they do not exist in PastAssociations.

Closes #377

BREAKING CHANGE: This change adds PlatformMetadata and modifies OperatorMetadata on the Metadata Spec.
It also removes the catalog merging workflow.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat(operators): adds changes to handle differential use-cases when processing IncludeConfigs

When processing catalog differentials, it is possible that modifications can
be made to the content that render the current IncludeConfig invalid. This change
adds an UpdateIncludeConfig method which will update the IncludeConfig used to
add new channels and packages and handle pruned starting bundles.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: removes check for adding the Cincinnati graph base image and always adds

To ensure the UBI image is never pruned before the creation of the of the
Cincinnati graph data image. The image must always be added if the graph
equals true.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: changes the way AssociationSets are pruned to avoid dropping Associations from the set

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: incorporates feedback from PR code review

Reduces code duplication in mirror.go

Adds comment to clarify nullLogger for containerd Registry

Adds a check when proccessing next bundles in case there
is not next bundle and the channel head does not exist. Unlikely,
but if truem results in a panic.

Changes catalog image build to remove the entire configs
directory from the base image before adding the FBC to avoid
possible FBC aggregation with previous yaml or json file.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>