fix: changes ConvertToAssociationSet methods to allow for shared child manifest (#406)
Currently, Associations contain multiple entries for one manifest that belong to more
than one image. These values are loaded into a map during the conversion which can
cause the child Association to end up with the wrong image data (e.g. Path/Type). Since the
combination of name(digest) and path(disk location) in the Association spec would be unique, that
is being used as the new key in the map so the image information is retained.

Related to PR #401
Fixes #402

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>