fix: adds handling for channel with different prefixes in cincinnati.go
Currently, the prefix for the target channels is being inferred from the source channel
when calculating Cincinatti graphs. This can cause an issue when The prefix of the channel
is different for the starting and target version. Now, the starting prefix is only
used when calculating upgrades through the channel until the target major and minor version
is reached.

Fixes #292

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>