fix(mapping): removes DockerClientDefaults from ctlgRef
When reading the catalog reference provided by the configuration,
the run method in operator.go was modifying the reference in a way
that resulted in the reference differing from the mapping. The DockerClientDefaults
call has been removed and a SetDefaults method has been added to TypedImage as a way to set
fields in a standard way on this type.

Closes #423

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>