chore(api-change): changes JSON field on ImagesetConfiguration spec to match the type names
In order to conform to Kubernetes API guidles, the Go field names should match the
JSON field names. Changed from the shortened version of the names to match the Go
field names and to line up with the examples in the documentation.

BREAKING CHANGE: Change JSON field names for Repositories, ImagePath, and
SkipDependencies

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>