fix(mapping): removes DockerClientDefaults from ctlgRef (#424)
* fix(mapping): removes DockerClientDefaults from ctlgRef

When reading the catalog reference provided by the configuration,
the run method in operator.go was modifying the reference in a way
that resulted in the reference differing from the mapping. The DockerClientDefaults
call has been removed and a SetDefaults method has been added to TypedImage as a way to set
fields in a standard way on this type.

Closes #423

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* test: adds end to end tests for full catalog mirror with digest

fix: create a helper function to get partial digest instead of manually
checking the string

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: fixes a comment in association_builder.go when referencing the default tag

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: adds a length check to the digest to avoid out of range on getPartialDigest

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: give numbers used in partial digest calculation a named variable.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>