feat: removes image blocking from bundle package and changes to regexp search
The image blocking isBlocked method needs to be moved from the bundle
package to the mirror package because it is only used by the
mirrorMapping method and bundle should be for imageset preparation.

This also changes the blocked images to use regular expression searching
to allow for more control over which images are blocked without the
requirement to specify a long list.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>