feat(catalogs): add catalog target name and tag (#432)
* feat: adds TargetName and TargetTag to Operator spec

BREAKING CHANGE: Extends the Operator type field in ImagesetConfigurationSpec
by adding field TargetName and TargetTag.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: removes requirement for full set to true with an include config

The IncludeAdditively workflow for processing catalogs has not been used and
was blocked with input validation at the configuration level. This lead to
confusing key combinations with package specifications and full being set to true.
IncludeAdditively is now always set to false.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* test: adds end to ends test for catalog with target information set

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: use and test catalog target names in validate.go

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix(manifests): modifies WriteCatalogSource method to account for name duplicates

Currently, duplicate repo names in a catalog image location will cause
catalog source manifest to be overitten. This changes finds names that
have already been used and adds an incremented integer as a suffix to the
manifest metadata name and filename if found.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: remove merge files with dead code in operator package

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>