feat(mirror): extends image blocking to allow the names to be specified as regular expressions (#436)
* feat: removes image blocking from bundle package and changes to regexp search

The image blocking isBlocked method needs to be moved from the bundle
package to the mirror package because it is only used by the
mirrorMapping method and bundle should be for imageset preparation.

This also changes the blocked images to use regular expression searching
to allow for more control over which images are blocked without the
requirement to specify a long list.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* test: adds additional test cases and benchmark to isBlocked function

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: adds some comments fixes and clarifications

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>