init workflow (#416)
* Omit empty minVersion and maxVersion

* Export GetVersion for future use

* Reword help message for --short

If the flag exists, it's true by default.
`--short=false` is still possible, but that's the same as not passing a
--short flag, so who would bother?

* wip: Add `oc-mirror init` command

The package is called initcmd because the Go compiler requires something called "init" to be a function.

WIP / not committed:
- reading the helm chart
- tests
- get catalog version and prompt for registry in parallel?

* Extract GetVersionMap to generic image.GetTagsFromImage

* Add ctx to initcmd.Run

Dead code, but no extra weight and maybe useful later

* Extract function

* Only ignore tags file, not tags dir

Preparing a future commit...

* init: Move block for cosmetics

* Explicitly depend on yaml.v2

The k8s yaml library is not able to properly serialize MapSlice, so
prepare to directly use yaml.v2

* init: Parameterify the catalog base for unit tests

Ugly magic string

* init: Remove extra \n from output

* init: Sort yaml keys

* init: Add tests

* init: Add README note for local builds

* init: Add fake v0.0.0- prefix to versionFromGit

For `make test-unit` to have something parseable without ldflags

* init: Update tests to run from make

* init: Replace magic context string with struct field

* init: Add tiny code review updates

* Fix misleading function name

GetTagsFromImage did processing on the tags before returning them.
The processing has been split out to GetVersionsFromImage

Co-authored-by: d10n <d10n@redhat.com>