feat(operator): allow ranges to be specified for package filtering (#421)
* chore(api): changes semver.Version field in IncludeConfig to strings

To allow for DeepCopy methods to be created for Metadata and ImageSetConfiguration
specs in the future the semver.Version fields were converted to string.

BREAKING CHANGE: Removes StartingVersion and StartingBundle from the spec
and adds MinVersion, MaxVersion, and MinBundle

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat(operators): adds version ranges and max version to IncludeConfig

Functionality is added to operator package filtering by allow a max
version to be specified. The original implementation is kept in place
for min version (starting version) and semver range strings are constructed
and passed as input to the declarative config.

Closes #302

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* build(deps): updates operator-registry dep back to official upstream

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* chore: incorporates changes for passing init command unit tests

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>