feat(release): expose architecture filter options for Platforms (#445)
* chore: exposes the filter-options flag on mirror and list commands

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* feat(release): adds the Architecture field to the Platform spec

Currently, platform release images are mirrored with "amd64"
as the default. The hidden filter-options flag has been removed from
'list updates' and the root 'mirror' command to allow the architectures
to be describe in the ImagesetConfigurationSpec since it only applies to
platform versions.

BREAKING CHANGE: Extends ImagesetConfigurationSpec, Platform specifically
to add the Architectures field.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>