fix(pruning): changes image planning for pruning operations to use the destination image repo (#444)
* fix(helm): changes the search method in helm.go to return all errors to the caller

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix(pruning): changes image planning for pruning operations to use the destination image repo

Currently, the source repo and destination repo for OCP releases are not the same.
This causes the image pruning planning operation to add an image in the wrong repository
to the mapping. The association path which defines the destination (in the archive or registry)
is used in planning with this change to determine the correct image repository information.

Fixes #447

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>