fix(operator): handles packages with multiple configured channels
In the current implementation, if a package is defined with
multiple channels in the current (user-defined) include config,
any channels without bundle information set will be recorded as is
if another channels has bundle information. This will result in a
"Version is empty" error on subsequent runs as the proper minimum
bundle version is never set before reconciliation.

This changes processed package level version configuration and
channel level version configuration seperately so new channels
will be processed properly.

Fixes #465

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>