fix(operator): adds context to error when bundles have invalid related images
Older catalogs may have bundles that contain related images with no actual
image information. The error that is thrown when this occur is not descriptive
enough to allow the user to take any kind of action. This change adds the bundle
and image information to allow the user to report back errors to the the publisher.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>