fix(pruning): allows continue-on-error to be used with pruning operations
Currently, if pruning errors are transport errors, only a warning is logged.
For users who want the pruning functionality, this can cause errors on runs
if the registry or authorization is not configured to handle deletion. Instead
of logging an error, an error is returned when pruning operations fail, but
if the error is a transport error, it can be ignored with continue-on-error.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>