fix(backend): pings the registry during metadata check to ensure validity
In the registry backend impl, 404 errors throw an ErrMetadataNotExist error
which allows operations to continue. For a registry that is not accessible,
this is an issue because the failure is delayed until the end when the
metadata publishing occurs. This changes adds a step to ping the registry
before determining the metadata does not exist.

Fixes #249

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>