Bug 2099899: fix(publish): modifies layer fetcher to gather repos into a map and u… (#492)
* fix(publish): modifies layer fetcher to gather repos into a map and uses assoc.Path instead of images

Currently, the image repo is being calculated each time a missing layer is identified.
Since the association set used is always the same, this is pretty inefficient. This change
creates a map with blobs as the keys and repo as the value. The first found repo will
be the value. This also changes the values from the image key to the assoc.Path values to account
for images being mapped to different repos (e.g. release images).

Fixes #491

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>

* fix: adds checks  to ensure repos can be found for association from mirror-to-mirror runs

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>