Bug 2029367: fix(config): Parse X.Y format minVersion and maxVersion
Error info is ambiguous when minVersion and maxVersion are wrong in the
ImageSetConfiguration, especially when using yaml.

The yaml unmarshaler parses unquoted X.Y values as a number, but users
intend that to be a semver version string.

The k8s yaml unmarshaler serializes to JSON and deserializes from JSON
in order to use the json marshaler's struct tags.

The json unmarshaler can't read a number into a string field, so a
custom UnmarshalJSON method is needed.

The custom json unmarshaling code would be simpler if MinVersion and
MaxVersion were extracted to a dedicated struct and inlined back into
ReleaseChannel, but there are a bunch of usages that would need to be
updated and I chose to prefer slightly-more-complicated unmarshaling
over clobbering history in a bunch of other places.

Fixes github issue #240