Replace tmp folder in current working directory by olm_artifacts folder under the oci mirroring destination.
olm_artifacts is within the FBC destination folder and is used  to untar the catalog contents,
in order to process it and prepare copy.
For each operator, a folder with the name of the operator will be created under artifactsPath (ctlcConfigsDir).
This fixes the comment left in https://coreos.slack.com/archives/C02JW6VCYS1/p1665437046724369?thread_ts=1665436944.265719&cid=C02JW6VCYS1