OCPBUGS-3414: Fix: fixes issues encountered by QE (#536)
* Move parseImageName and trimProtocol  to image pkg

* OCI Case: add OriginalRef to TypedImage for ICSP generation

* Use relatedImage ref instead of name

This was initially used so that the tree of images (and their blobs) is flat under oc-mirror-workspace.
The intent was to gain on performance by not pulling several times same images when they are referenced in several catalogs.
This results in errors in generated ICSP as the namespace of the image is lost

* Fix the destination of the catalog image for ICSP generation