fix the graph-data container build logic wrt untar
we want the new graph-data image to be compatible with
both dkrv2 plugin of cincinnati and when used as an
init container with Update Service.

The old logic extracted the tarball when the container
was executed. This caused the graph-data directory to not
exist when used with the dkrv2 plugin as we wont execute
the container.

with this change, we untar the tarball before creating the
image, so the directory exists even not executed.