[graph-data] use new cincinnati endpoint to download tarball (#552)
* use api.openshift.com uri to download graph-data

with /graph-data endpoint in cincinnati, we should
be able to download the graph-data tarball directly
from UpdateService(cincinnati) instead of reaching
to github.com

* fix the graph-data container build logic wrt untar

we want the new graph-data image to be compatible with
both dkrv2 plugin of cincinnati and when used as an
init container with Update Service.

The old logic extracted the tarball when the container
was executed. This caused the graph-data directory to not
exist when used with the dkrv2 plugin as we wont execute
the container.

with this change, we untar the tarball before creating the
image, so the directory exists even not executed.