First pass for multi architecture catalogs
- Numerous changes to read/write single and multi arch images. Each code
path has to handle the parent/child hierarchy of indexes, usually via
recursion.
- Made the operator rendering code "group together" the
DeclarativeConfig and IncludeConfig and keep track of them on a per
platform basis. Introduced this and other new data structures to pass
between the various functions in operator.go.
- some of the changes in fbc_operators.go and fbc_operators_test.go are
bare minimum changes to make the code compile. This code is likely to
have changed significantly in other PR's.
- some items are left todo... areas like GetUniqueName will likely need
to change based on other PRs. This has impacts in store.go.
- update bunch of test cases and introduce new tests
- added docs as to clarify pieces as I understood them, and some of
these changes are unrelated to the core purpose of this PR
- updated go.mod and vendor