fix extract dir for cincinnati-graph-data container
we're changing the dir for extracting cincinnati-graph-data
as we're mounting an empty volume to /var/lib/cincinnati/graph-data
on the operator. this results in the dir getting overrided with
an empty dir.

Solution:
Extract cincinnati-graph-data to a different directory and copy the
contents to the required directory after the container is mounted
by introducing an init command for the container.