fix broken test cases
- TestGetCatalogMetadataByPlatform was becoming too complex to maintain
so it was re-written to just start a fake server, push content to it,
and then run the tests that talk to this server
- TestGetCatalogMetadataByPlatformFromOCILayout no longer uses the
"hello" test image since it does not contain an actual catalog. Also
added a single image test case to this test
- Add a single "testonly" image to use instead of
pkg/cli/mirror/testdata/artifacts/rhop-ctlg-oci which is actually broken
and therefore can't be pushed to a registry.