fix issue with image builder
An OCI layout can reference a manifest list either directly or from an
"indirect index reference" found in the blobs. This "indirect index
reference" is NOT possible for a docker registry. This fix makes
the image builder "discover" the right index to push, so that you get
the correct content pushed into the remote registry. See code comments
for details.