fix validateMapping OCPBUGS-13962 (#640)
The validateMapping function would output warning messages when it did
not need to. This fix attempts to find either an exact match or a
partial match based on the image name (i.e. the last component of the
image reference) as well as name:tag, name@digest, or name:tag@digest
forms. Added test case TestValidateMapping to check the log output for
the corret messages. The klog test package was also included in the unit
tests, so vendor directory was updated too.