archive: Moving context argument from constructors to methods
This is done to anticipate instanciation of Archiver in the
Executor.go.
In fact,  the cobra command context will only be available in the Run
method. The calls to `archive.NewMirrorArchive` which are expected in
methods Complete and PrepareComplete of the executor happen before
Run, and therefore the context of the command isn't created yet.
Therefore, we are moving the context argument from `NewImageBlobGatherer`
and `NewMirrorArchive` into the methods `GatherBlobs` and `BuildArchive`
respectively.