Build archive in MirrorToDisk workflow
Adds the creation of an implementation of interface Archiver in the
`Complete` method.

Adds the call to Archiver.BuildArchive in `Run` for MirrorToDisk mode
after the Batch workers have finished.
Before building the archive, the local storage registry is stopped, so
that we ensure that the cache folder is in a consistent state before
we start archiving.