CFE-977,CFE-991: archive generation at the end of MirrorToDisk workflow (#739)
* Add v2/tests/results/cover* to .gitignore

* Rename Dir to WorkingDir in type GlobalOptions

* archive: Moving context argument from constructors to methods

This is done to anticipate instanciation of Archiver in the
Executor.go.
In fact,  the cobra command context will only be available in the Run
method. The calls to `archive.NewMirrorArchive` which are expected in
methods Complete and PrepareComplete of the executor happen before
Run, and therefore the context of the command isn't created yet.
Therefore, we are moving the context argument from `NewImageBlobGatherer`
and `NewMirrorArchive` into the methods `GatherBlobs` and `BuildArchive`
respectively.

* Build archive in MirrorToDisk workflow

Adds the creation of an implementation of interface Archiver in the
`Complete` method.

Adds the call to Archiver.BuildArchive in `Run` for MirrorToDisk mode
after the Batch workers have finished.
Before building the archive, the local storage registry is stopped, so
that we ensure that the cache folder is in a consistent state before
we start archiving.

* Update unit test for MirrorToDisk use case