OCPBUGS-23327: Fix MirrorToDisk of oci catalogs in hidden folders
When constructing the mapping (and mapping.txt) for OCI catalogs, and calculating
the destination for mirroring of related images of the catalog,  we don't add
(nor later remove) the catalog path from the destination.
This way, the destination obtained for a relatedImage is (supposed to be) same as for
MirrorToDisk workflow with registry based catalogs.
This fixes the issue in OCPBUGS-23327, since the . character is no longer part of the
reference to the destination related image.