OCPBUGS-44818: Fix mirroring catalogs with full:true (#977)
Also fixes the spinner staying stuck when errors happen during collection of operators.
Introduces a `ToRebuild`(bool) to the `CatalogFilterResult` struct, which is used to handover catalogs between the collection phase and the rebuild phase.
With this flag, the catalogs in full:true (no need for rebuild) will be skipped at the rebuild phase.
A benefit from this commit is to save time by not rebuilding catalogs that are already rebuilt in a previous mirroring.

Co-authored-by: sherine-k <skhoury@redhat.com>