v2/operator: skip failed operators
If we fail to collect an operator, save the error but keep collecting
the other operators on the list. At the end, we return all errors found.

Further work is need to not stop at collection and proceed to the copy
but still return non-zero exit status at the end of oc-mirror execution.