OCPBUGS-50680: v2/operator: refactor operator FilteredCollector (#1093)
* v2/operator/common: dedup code for catalogDigest

I have also added a unit test for `cachedCatalog`.

* v2/api: prepare CatalogFilterResult for refactor

We will need this information so we can refactor the OperatorCollector.

* v2/operator: refactor FilteredCollector

Instead of an endless chain of if-else, let's separate blocks into
function according to their actions.

This makes the collector easier to read, remove error handling
duplication (like aborting the spinner), and it will make easier to skip
operators but return a list of failed operators at the end.

* v2/operator: skip failed operators

If we fail to collect an operator, save the error but keep collecting
the other operators on the list. At the end, we return all errors found.

Further work is need to not stop at collection and proceed to the copy
but still return non-zero exit status at the end of oc-mirror execution.

* v2/operator: copy OCI operator on any errors

Before we were just copying the operator contents to the working dir if
the `stat` call returned `ErrNotExist`. To be safe, we should try to
copy on any errors.

* v2/operator: fix linting issues

We'll have to postpone fixing the cyclop issues until we are able to
refactor the code even further.