Remove unused implementation CatalogBuildah for catalog rebuild
This is needed as part of this fix, because catalog-filter v0.0.4 contains an update of go-git library (fixing vulnerability).
This change in go-git library is forcing an update of github.com/cyphar/filepath-securejoin v0.4.1.
This library is also used by runc, (dependency from buildah), and would cause the following issue:
```
/go/pkg/mod/github.com/opencontainers/runc@v1.2.1/libcontainer/utils/utils_unix.go:337:56: cannot use int(mode) (value of type int) as os.FileMode value in argument to securejoin.MkdirAllHandle
```