OCPBUGS-49880: cli: return distinct error codes depending on failure (#1062)
* v2/cli: move mirror time calculation to parent function

Instead of duplicating the logic in each run mode.

* v2/cli: small coding style changes

* Remove unnecessary newlines
* Normal comments should have a space after `//` to distinguish them
  from comment directives.

* v2/cli: simplify code flow by returning early from dry-run

* v2/batch: remove `copiedImages` package variable

As a package variable there is only one instance of it which is not
threadsafe. Instead we want variables to be returned for each run of the
concurrent_chan_worker.

* v2/cli: simplify batch runner error check

The new concurrent_chan_worker always returns `SafeError` so there is no
point anymore on the distinction between `SafeError` and `UnsafeError`.

* v2/batch: remove concurrent_worker

It's been completely replaced by concurrent_chan_worker. Removing this
will avoid useless changes to concurrent_worker anytime we want to make
changes to the batch interface.

* v2/internal: use time.Since() instead of time.Now().Sub()

* v2/release: move big if-else blocks to their own functions

By being in their own functions, it's clear to the reader that there is
no variable/functionality in common between the 2 modes. This should
make it easier to reason about each.

* v2/common: add exit return codes consts

These will be useful to return exit codes according to the type of
errors found.

* v2/cli: introduce a `CodeExiter` interface

This interface can be implemented by errors causing oc-mirror to exit
with the specified exit code.

* v2/cli: return exit codes according to collection errors.

Introduce a new `CollectionError` type that implements the `CodeExiter`
interface so we can return appropriate error codes from collection
errors.collection errors.collection errors.collection errors.

* OCPBUGS-49880: v2/cli: report non-zero exit in case of errors

If there are collection or copying errors, `oc-mirror` will now exit
with an exit code other than 0.

* v2/release: lower cyclomatic complexity of collector.

Move independent blocks of code to functions so they are easier to
reason about.

* v2/cli: remove unused create cache dir code

The filtered_collector is the default now, so doing as the comment says.

* v2/batch: fix gosec and ireturn linter issues

`MaxGoroutines` is not supposed to be negative, so let's use `uint`
instead of `int` and avoid the `uint -> int` conversion altogether.