Disable go workspaces and run v2 as separate binary
Due to the way go workspaces work, we can not upgrade/downgrade packages
without affecting both oc-mirror v1 and v2. Removing the workspace
allows us to have distinct dependency versions thus making it possible
to have separate development cadences for v1 and v2.

To make it possible for users to still invoke --v1 or --v2 from the same
binary, we use the `//go:embed` directive to embed the v2 binary into
the v1 oc-mirror, extract and execute the binary at runtime. It incurs a
small time penalty but it has the same effect as importing the v2 module
into v1 without the downsides.