OCPBUGS-54220: v2/imgbuilder: use --authfile when set (#1129)
Right now the go-containerregistry lib is using the default
docker+podman locations. If oc-mirror is run with `--authfile` pointing
to a custom location, it will be used for other images except the graph
image.

This change adds a custom keychain with the contents of --authfile so
it is picked up by go-containerregistry.

Fixes #1122