OCPBUGS-56185: [release-4.19] v2: bump containers/common dep to 0.63.0 (#1148)
* cli/mirror: lint fixes

* cmd/oc-mirror: small refactoring to separate v1/v2

We shouldn't be importing the v2 module into v1 (logically wrong).
Having them as separate modules will allow us to bump their dependencies
independently from one another, enabling faster pace development in the
v2 code.

This change separates v1 and v2 executions into separate blocks. This
will make it easier to run v2 as a separate binary in the next commit.

* v2/cli: do not os.Exit from the cobra command

We want to have a chance to finish tasks in case oc-mirror execution
fails, e.g., writing out the cpu and mem prof files. Doing `os.Exit`
directly means that not even deferred functions can run, so we just want
to do that at the topmost level of `main`.

* Disable go workspaces and run v2 as separate binary

Due to the way go workspaces work, we can not upgrade/downgrade packages
without affecting both oc-mirror v1 and v2. Removing the workspace
allows us to have distinct dependency versions thus making it possible
to have separate development cadences for v1 and v2.

To make it possible for users to still invoke --v1 or --v2 from the same
binary, we use the `//go:embed` directive to embed the v2 binary into
the v1 oc-mirror, extract and execute the binary at runtime. It incurs a
small time penalty but it has the same effect as importing the v2 module
into v1 without the downsides.

* v2/main: fix linting issues

This code was running from the v1 module, where we don't run the linter,
so these issues were never addressed before.

* Makefile: clean up and separate v2 where possible

* Removed unused definitions
* Consolidate `DISABLE_CGO` and `CGO_ENABLED` into the latter.
* Also updated `openshift/build-machinery-go` to latest version.

* OCPBUGS-54587: v2: bump containers/common dep to 0.63.0

This updates brings in a fix to retry the image copy operation in case
of 50[234] errors which happen quite often with quay.io.