OCPBUGS-59235: [release-4.18] v2/archive: error d2m when no tar files found (#1212)
* OCPBUGS-58193: v2/archive: error d2m when no tar files found

If we don't error out right away, the error message returned gives no
clue about the actual reason for the failure:
```
2025/06/27 10:21:45  [ERROR]  : [Executor] collection error: collect catalog "registry.redhat.io/redhat/redhat-operator-index:v4.18": get manifest: error when creating a new image source: reading manifest v4.18 in localhost:55000/redhat/redhat-operator-index: manifest unknown
```

* v2/cli: fix unit tests

Running d2m with a working dir that doesn't contain any tar files should
result in an error.

Also replaced some instances of
```
if err != nil {
    t.Fatalf("should not fail")
}
```
with
```
assert.NoError(t, err)
```
so we get the error message in the test failure output.