OCPBUGS-59646: cmd: unpack v2 to $HOME instead of /tmp (#1220)
It seems it's quite common for users to run oc-mirror in a system with
the /tmp dir mounted `noexec` for STIG compliance. In that case,
oc-mirror v2 would silently fail because we were not properly printing
the error output and the binary could not be executed.

Even though it's possible to workaround the issue by temporarily
overriding the default tmp dir with the `TMPDIR` envvar, let's change
the target directory for unpacking v2 to `$HOME/.oc-mirror` so the
binary works in STIG-compliant systems out-of-the-box.