OCPBUGS-59646: cmd: print errors when v2 cannot be run
We were printing only the errors returned by v2 after it was executed
but we were not printing the errors during the v2 binary setup. This
would result in no console output and oc-mirror just failing silently.

This change makes sure that we both output the error and return a
non-zero status code.