OCPBUGS-59646: revert #1220 and print a warning instead (#1228)
* Revert "OCPBUGS-59646: cmd: unpack v2 to $HOME instead of /tmp (#1220)"

This reverts commit 2533318e9933288543704046c3e4bf9b0975624f.

It seems /home must also be mounted `noexec` [1]. So let's go back to
extracting to /tmp since overriding it with the `TMPDIR` envvar is less
disruptive than overriding `HOME`.

[1] https://stigviewer.com/stigs/red_hat_enterprise_linux_9/2025-02-27/finding/V-257852

* OCPBUGS-59646: cmd: print errors when v2 cannot be run

We were printing only the errors returned by v2 after it was executed
but we were not printing the errors during the v2 binary setup. This
would result in no console output and oc-mirror just failing silently.

This change makes sure that we both output the error and return a
non-zero status code.

* OCPBUGS-59646: cmd: print a warning if TMPDIR is `noexec`

Instruct users to set `TMPDIR` to a filesystem with `exec` permissions
so that oc-mirror v2 can be run.