[release-4.19] OCPBUGS-59760: warn if tmp is `noexec` (#1234)
* OCPBUGS-59760: cmd: print errors when v2 cannot be run

We were printing only the errors returned by v2 after it was executed
but we were not printing the errors during the v2 binary setup. This
would result in no console output and oc-mirror just failing silently.

This change makes sure that we both output the error and return a
non-zero status code.

* OCPBUGS-59760: cmd: print a warning if TMPDIR is `noexec`

Instruct users to set `TMPDIR` to a filesystem with `exec` permissions
so that oc-mirror v2 can be run.