ocpbugs-60999: fix issue when using full true in m2m (#1261)
Before this fix during m2m when using full: true, the dispatch func was
trying to retrieve a rebuilt catalog image from cache. This behavior was
wrong since full: true does not rebuild the catalog.

This PR fixes this behavior, it gets the original catalog image in case no RebuiltTag was found for the catalog image.

Co-authored-by: Alex Guidi <aguidi@redhat.com>